rename files and update make pattern
