+++ fs/checkpoint.c
@@ -1,11 +1,11 @@
-#include "jessfs.h"
+#include "virfs.h"
 
 #include "ifos.h"
 #include "threading.h"
 
 #include "runtime.h"
-#include "jessos.h"
-#include "jesspro.h"
+#include "refos.h"
+#include "order.h"
 
 /* fs server object, def in jessfs.c */
 extern struct fs_server __fs_server;
