+++ ldr/entry.c
@@ -1,10 +1,10 @@
 #include "logger.h"
 
-#include "jessnet.h"
-#include "jessfs.h"
-#include "jessldr.h"
-#include "jessos.h"
-#include "jesspro.h"
+#include "interact.h"
+#include "virfs.h"
+#include "ldr.h"
+#include "refos.h"
+#include "order.h"
 #include "conf.h"
 
 #include "args.h"
