+++ ldr/ldr.c
@@ -1,11 +1,11 @@
-#include "jessldr.h"
+#include "ldr.h"
 
 #include "ifos.h"
 #include "threading.h"
 
 #include "runtime.h"
-#include "jessos.h"
-#include "jessfs.h"
+#include "refos.h"
+#include "virfs.h"
 
 static const char *__jessldr_scope_level_name[] = { "dir", "archive", "logic", "array", "item", "filed"};
 // static const char __jessldr_invalid_scope_name_component_character[] = {
