+++ ldr/ldr.h
