+++ pro/variable.c
@@ -1,10 +1,10 @@
-#include "jesspro.h"
+#include "order.h"
 
 #include "runtime.h"
-#include "jessfs.h"
-#include "jessldr.h"
-#include "jessnet.h"
-#include "jessos.h"
+#include "virfs.h"
+#include "ldr.h"
+#include "interact.h"
+#include "refos.h"
 
 static void pro_scan_current_scope_level(const struct avltree_node_t *tree, JRP *jrp)
 {
