+++ runtime/runtime.c
@@ -1,9 +1,9 @@
 #include "runtime.h"
 
 #include "request.h"
-#include "jessldr.h"
-#include "jessfs.h"
-#include "jessos.h"
+#include "ldr.h"
+#include "virfs.h"
+#include "refos.h"
 
 #if _WIN32
 	/* #pragma warning(disable:) */
